符号付きと符号なしの比較を避けるようにする

+++ graphic/scenes/proto.cpp
@@ -5,9 +5,9 @@ namespace mihajong_graphic {
 
 int Scene::whichRegion(int X, int Y) {
 	int regionNum = -1;
-	for (int i = 0; i < regions.size(); i++)
+	for (unsigned int i = 0; i < regions.size(); i++)
 		if ((X >= regions[i].Left) &&(Y >= regions[i].Top) && (X <= regions[i].Right) && (Y <= regions[i].Bottom))
-			regionNum = i;
+			regionNum = static_cast<int>(i);
 	return regionNum;
 }
